% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cetesb_param_ids.R
\docType{data}
\name{cetesb_param_ids}
\alias{cetesb_param_ids}
\title{Ids for the CETESB parameters.}
\format{A data frame with 20 rows and 3 variables:
\describe{
\item{id}{Id of the parameter.}
\item{param_abbrev}{Abbreviation of the parameter name.}
\item{param}{Parameter name.}
}}
\source{
\url{http://qualar.cetesb.sp.gov.br/}
}
\usage{
cetesb_param_ids
}
\description{
A dataset containing the ids for every CETESB parameter.
}
\keyword{datasets}
