% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cetesb_station_ids.R
\docType{data}
\name{cetesb_station_ids}
\alias{cetesb_station_ids}
\title{Ids for the CETESB stations.}
\format{
A data frame with 62 rows and 3 variables:
\describe{
\item{id}{Id of the station.}
\item{stationname}{Name of the station.}
\item{initial_date}{Date when the station started to operate,
in format dd/mm/aaaa.}
\item{BEN}{Indicates if the parameter BEN is measured or not in the station.}
\item{CO}{Indicates if the parameter CO is measured or not in the station.}
\item{ERT}{Indicates if the parameter ERT is measured or not in the station.}
\item{MP10}{Indicates if the parameter MP10 is measured or not in the station.}
\item{MP2.5}{Indicates if the parameter MP2.5 is measured or not in the station.}
\item{NO}{Indicates if the parameter NO is measured or not in the station.}
\item{NO2}{Indicates if the parameter NO2 is measured or not in the station.}
\item{Nox}{Indicates if the parameter Nox is measured or not in the station.}
\item{O3}{Indicates if the parameter O3 is measured or not in the station.}
\item{SO2}{Indicates if the parameter SO2 is measured or not in the station.}
\item{TOL}{Indicates if the parameter TOL is measured or not in the station.}
\item{DV}{Indicates if the parameter DV is measured or not in the station.}
\item{DVG}{Indicates if the parameter DVG is measured or not in the station.}
\item{PRESS}{Indicates if the parameter PRESS is measured or not in the station.}
\item{RADG}{Indicates if the parameter RADG is measured or not in the station.}
\item{RADUV}{Indicates if the parameter RADUV is measured or not in the station.}
\item{TEMP}{Indicates if the parameter TEMP is measured or not in the station.}
\item{UR}{Indicates if the parameter UR is measured or not in the station.}
\item{VV}{Indicates if the parameter VV is measured or not in the station.}
\item{address}{Address of the station.}
\item{lat}{Latitude of the station.}
\item{long}{Longitude of the station.}
}
}
\source{
\url{http://qualar.cetesb.sp.gov.br/}
}
\usage{
cetesb_station_ids
}
\description{
A dataset containing the ids for every CETESB station.
}
\keyword{datasets}
