% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_air_pollution.R
\name{get_air_pollution}
\alias{get_air_pollution}
\title{Returns a map depicting varying pollution levels for a specified location}
\usage{
get_air_pollution(lat, lon, api_key, fig_title)
}
\arguments{
\item{lat}{Latitude for a given location as a double}

\item{lon}{Longitude for a given location as a double}

\item{api_key}{The API key for OpenWeather as a string}

\item{fig_title}{The title of the returned ggplot as a string}
}
\value{
ggplot object
}
\description{
The function makes an API request to the OpenWeather Air Pollution API
and fetches pollution data for a given location, then transforms
the returned JSON object from the request into a ggplot chart.
}
\examples{
get_pollution_history(49.28, 123.12, "APIKEY_example", "example_title")
}
