% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pollution_history.R
\name{get_pollution_history}
\alias{get_pollution_history}
\title{Returns a dataframe of pollution history for a location between
a specified date range.}
\usage{
get_pollution_history(start_date, end_date, lat, lon, api_key)
}
\arguments{
\item{start_date}{The start date as a UNIX timestamp, e.g. 1606488670}

\item{end_date}{The end date as a UNIX timestamp, e.g. 1606747870}

\item{lat}{Latitude coordinate for a given location as a double}

\item{lon}{Longitude coordinate for a given location as a double}

\item{api_key}{The API key for OpenWeather as a string}
}
\value{
tibble with data containing pollution history
}
\description{
Given a specified date range, the function makes an API request
to the OpenWeather Air Pollution API and fetches historic
pollution data for a given location.
}
\details{
Note: Historical data is accessible from 27th November 2020
}
\examples{
get_pollution_history(1606488670, 1606747870, 49.28, 123.12, "APIKEY_example")
}
