% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infoaguas_clean_dataset.R
\name{infoaguas_clean_dataset}
\alias{infoaguas_clean_dataset}
\title{Clean the data returned by the Infoaguas system}
\usage{
infoaguas_clean_dataset(df_results)
}
\arguments{
\item{df_results}{tibble with the data returned by the Infoaguas system.}
}
\value{
A cleaned tibble with the data returned by the Infoaguas system.
}
\description{
Clean the data returned by the Infoaguas system
}
