% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infoaguas-login.R
\name{infoaguas_login}
\alias{infoaguas_login}
\title{Login at Infoaguas CETESB.}
\usage{
infoaguas_login(login, password)
}
\arguments{
\item{login}{Character. Login email.}

\item{password}{Character. Password}
}
\value{
Message wheter the login succeeded or not.
}
\description{
Register at: https://sistemainfoaguas.cetesb.sp.gov.br/Login/Index
}
