% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infoaguas-scraper.R
\name{infoaguas_scraper}
\alias{infoaguas_scraper}
\title{Scraper Infoaguas CETESB}
\usage{
infoaguas_scraper(
  station,
  start = NULL,
  end = NULL,
  login,
  password,
  file = NULL,
  clean = FALSE
)
}
\arguments{
\item{station}{A numeric value indicating the station id from
where you wish to get the data.
See \link{infoaguas_station_ids}.}

\item{start}{A string in the format "dd/mm/aaaa" representing
the initial day for the data selection. If NULL (default) the date of
the start of the operation of the station will be used.}

\item{end}{A string in the format "dd/mm/aaaa" representing  the
final day for the data selection. If NULL (default) the date of the
search will be used.}

\item{login}{A string with your login on Infoaguas system.}

\item{password}{A string with your passoword on Infoaguas system.}

\item{file}{A string containing a path to a .rds file where the data will be
written. If NULL (default) the data will only be returned.}

\item{clean}{Logical. If TRUE, the data will be cleaned.
If FALSE (default) the original data willbe returned.}
}
\value{
A tibble with the data returned by the Infoaguas system.
}
\description{
Scraper Infoaguas CETESB
}
\examples{
\dontrun{

# Data from the station BILL02100, in the Billings reservoir, in 2020.

infoaguas_scraper(
  station = 146,
  start = "01/01/2020",
  end = "31/12/2020",
  login = "login",
  password = "password",
  file = "146_billings.Rds",
  clean = TRUE
)
}
}
