% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPollution.R
\docType{package}
\name{RPollution}
\alias{RPollution}
\alias{RPollution-package}
\title{\code{RPollution} package}
\description{
Functions for air pollution analysis
}
\details{
See the README
on \href{https://github.com/williamorim/Rpollution#readme}{GitHub}
}
