% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infoaguas-login.R
\name{infoaguas_login}
\alias{infoaguas_login}
\title{Login at Infoaguas CETESB.}
\usage{
infoaguas_login(login, password)
}
\arguments{
\item{login}{A string with your login on Infoaguas system.}

\item{password}{A string with your passoword on Infoaguas system.}
}
\value{
Message whether the login succeeded or not.
}
\description{
Register at: https://sistemainfoaguas.cetesb.sp.gov.br/Login/Index
}
