% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPollution.R
\docType{package}
\name{Rpollution}
\alias{Rpollution}
\title{\code{Rpollution} package}
\description{
Functions for air pollution analysis
}
\details{
See the README
on \href{https://github.com/williamorim/Rpollution#readme}{GitHub}
}
