% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper_cetesb.R
\name{scraper_cetesb}
\alias{scraper_cetesb}
\title{Scraper CETESB}
\usage{
scraper_cetesb(station, parameter, start, end, login, password,
  safe = FALSE, file = NULL)
}
\arguments{
\item{station}{A numeric value indicating the station id from
where you wish to get the data.
See \link[Rpollution:cetesb_station_ids]{Rpollution::cetesb_station_ids}.}

\item{start}{A string in the format "dd/mm/aaaa" representing
the initial day for the data selection.}

\item{end}{A string in the format "dd/mm/aaaa" representing  the
final day for the data selection.}

\item{login}{A string with your login on Qualar system.}

\item{password}{A string with your passoword on Qualar system.}

\item{safe}{If FALSE (default) the function will return a error if it can't
extract the requested data.}

\item{file}{A string containing a path to a .rds file where the data will be
written. If NULL (default) the data will be return.}

\item{parameters}{A numeric vector of maximum length equal to 3 indicating
the ids of the parameters you want to get the data from.
See \link[Rpollution:cetesb_param_ids]{Rpollution::cetesb_param_ids}.}
}
\value{
A tibble with the data returned by the Qualar system.
}
\description{
Scraper CETESB
}
\examples{
\dontrun{

# Ozone for 'Dom Pedro II' station for January 2018.

scraper_cetesb(station = 72, parameter = 63,
               start = "01/01/2018", end = "31/01/2018",
               login = "login", password = "password")
}
}
